wm: fix instanceof usage
`instanceof` is an operator that quite frankly should not be used at all.
However, in lieu of replacement, we fix these instances to not cause
compiler complaints. You cannot inline define variables with the
object of an `instanceof` invocation because you have an instance
of a class, not of a specific class instance.

Change-Id: I48f6529bc85e253d3ffa9bc96fdf68205ea60f57