Biometrics: Hook up support for halHandlesDisplayTouches prop
Fingerprint AIDL interface supports boolean type sensor prop
halHandlesDisplayTouches, that defines whether the HAL is responsible
for detecting and processing of display touches. If the value is false,
the framework was supposed to be responsible for handling the display
touch events and passing them down to the HAL by using ISession#onPointerDown
and ISession#onPointerUp. If the value is true, the framework was not
supposed to notify the HAL about touch events.

Currently, Android framework is not aware of this prop, so hook it up
for the expected outcome.

Change-Id: I56166950c6ef7cbb0c4ac2c0b7998e4a909d4af9