SystemUI: Fixup center clock position
When clock position set to center, after a few minutes, there is an
empty space on the left side where the clock is on the left. This is
because it only checks the clock for the right position. So we make it
checking the clock in the center position too.

Change-Id: I4ed24f96106346ead9c12fa3fcab6e87387a552d