Track soft-rejected mic ops
See the frameworks/av change with this same Change-Id for more info.
This change will not work properly without that one, and vice versa.

Prior to this change, the startRecording and finishRecording calls in
our same-Change-Id counterpart in fw/av convey the following info from
our side, based on return value:
1. PERMISSION_GRANTED: Op started, app ops count increased.
2. PERMISSION_HARD_DENIED: Op not started, count not increased.
3. PERMISSION_SOFT_DENIED: Op maybe started, ops count maybe
   increased; there is no way to know.

This change aims to make the SOFT_DENIED case reliably increase the
ops counter, should an event already be present. Predictable behavior
is necessary to keep these states in sync. (The fact that a first-time
soft-denial rejection does not increase a counter is an unresolved
quirk necessary to behave as CTS tests expect.)

Test: atest CtsMediaAudioPermissionTestCases
Issue: calyxos#3123
Change-Id: I1a67234d8809898d47ab1845687f249b32831546